Rename the directory from ContentModerator to Crew
