+++ app/Http/Livewire/RoomAndTimeslotSelector.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Livewire;
 
-use App\Http\Controllers\ContentModerator\ScheduleController;
+use App\Http\Controllers\Crew\ScheduleController;
 use App\Models\DefaultPresentation;
 use App\Models\Room;
 use App\Models\Timeslot;
